Added a helper method to save data to a CSV
Added the write_csv function, which mirrors read_csv. You pass it a file
path and a dictionary of dictionaries. It determines the field names and
writes all of the current data in memory back out to the specified CSV
file.

Should be able to use this for attendees as well as any other dictionary
of dictionaries.