Use Python's UUID library to generate IDs instead
This fixes a previous bug where new IDs were created based on the number
of events. Dunno if this is the most intuitive thing for students, but
it's generally better not to use sequential IDs for things we might one
day want to wrap some security around. (Remember that Parler hack?)