Added ability to read a config file
Set up a config_defaults.py file, which contains the keys and default
values. For each "deployment", a config.py file should be made from a
copy of config_defaults.py, and then edited to contain the actual values
to use.

The config file should contain things that should not be stored in Git,
such as database credentials and the secret key.

The .gitignore file was also updated to ignore config.py.