Added routes and templates to handle attendees
Created new routes and a new template for adding or editing an attendee.
Updated event.html to show attendees for an event, and include an "add
attendee" button.

Also needed to udpate database.py to add a function for retrieving a
single attendee's info for editing. Fixed a type-o as well: the function
 "update_attendee" was misspelled as "update_attenddee".