Added ability to edit an event
I meant to do this before adding attendees, but forgot!!

Renamed the function create_event to edit_event so that it can be used
to handle inserts and updates, as much of the code is the same.

Renamed create_event.html to event_form.html so that it can be used for
inserts and updates too.

Updated event.html to add the edit link and updated events.html to
reflect the change in function name.